Throw an exception in IM_ASSERT to get a Python backtrace on ImGui errors (#3)
Pretty ugly.  The app will be in somewhat of an invalid
state on IM_ASSERT so there's probably not much point in trying
to continue when this happens.  But it's helpful for the programmer
to get a backtrace and some extra information on what went wrong
instead of just crashing or segfaulting.