Some fixes to function overloads
All overloads should now be output to docs.

However, not all overload docstrings are there, as it's
hard to match the comments between Python overloads and C++
ImGui overloads.  So make a cheap approximation by
considering only the first function overload comment from imgui.h
and use that, and provide no docs for the rest.

Probably reasonably ok, but a bit messy.