Wrap DrawList's in the Python wrapper for more fine-grained control of object lifetimes
This is needed for DrawList.add_callback calls which must be designed so that
the Python callables are kept alive via object references and don't get
deallocated before the next `imgui.new_frame()` is called.

The DrawList getters for the Context were changed from `nb::rv_policy::reference_internal`
to `nb::rv_policy::reference` to make the refcounts testing make more sense.  Sure
it's not safe if you hold on to a drawlist handle until after you've destroyed the ImGui context
and then try to call its methods, but that's not a very sensible thing to do.

for #22